[3276] Minor refactor
