+++ stream-chat-android-ui-components/api/stream-chat-android-ui-components.api
@@ -1737,10 +1737,6 @@ public abstract interface class io/getstream/chat/android/ui/message/list/Messag
 	public abstract fun onGiphySend (Lio/getstream/chat/android/client/models/Message;Lcom/getstream/sdk/chat/enums/GiphyAction;)V
 }
 
-public abstract interface class io/getstream/chat/android/ui/message/list/MessageListView$GroupedMessageListItemPredicate {
-	public abstract fun isGroupedWithNextMessage (Lcom/getstream/sdk/chat/adapter/MessageListItem$MessageItem;)Z
-}
-
 public abstract interface class io/getstream/chat/android/ui/message/list/MessageListView$LastMessageReadHandler {
 	public abstract fun onLastMessageRead ()V
 }
