+++ stream-chat-android-ui-components/src/main/kotlin/io/getstream/chat/android/ui/message/list/adapter/view/internal/FootnoteView.kt
@@ -80,6 +80,9 @@ internal class FootnoteView : ConstraintLayout {
         threadsFootnote.root.isVisible = false
     }
 
+    /**
+     * Hides the message footnote.
+     */
     fun hideSimpleFootnote() {
         footnote.root.isVisible = false
         threadsFootnote.root.isVisible = false
